# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CodeArena API",
    "description": "Production API specification for CodeArena — Online Judge & Secure Code Execution Platform.",
    "version": "1.0.0",
    "contact": {
      "name": "CodeArena Engineering"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Environment Server"
    }
  ],
...
  "paths": {
...
    "/api/v1/problems/{problemId}/test-cases": {
      "get": {
        "tags": ["Test Cases"],
        "summary": "Get test cases for a problem",
        "description": "Retrieves test cases for a problem with database-level role isolation. Normal users (USER role) receive ONLY active PUBLIC test cases. Administrators (ADMIN role) receive all test cases (both PUBLIC and HIDDEN).",
        "operationId": "getTestCasesForProblem",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "problemId",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the problem",
            "schema": {
              "type": "string",
              "example": "65f0123456789abcdef01235"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Accessible test cases for the problem",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "testCases": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TestCase"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "post": {
        "tags": ["Test Cases"],
        "summary": "Create a test case for a problem",
        "description": "Creates a new test case associated with the problem specified in the path. Restricted strictly to administrators (ADMIN role). Records administrative audit log.",
        "operationId": "createTestCaseForProblem",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "problemId",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the problem",
            "schema": {
              "type": "string",
              "example": "65f0123456789abcdef01235"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestCaseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Test case created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "testCase": {
                          "$ref": "#/components/schemas/TestCase"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/api/v1/test-cases/{testCaseId}": {
      "get": {
        "tags": ["Test Cases"],
        "summary": "Get test case by ID",
        "description": "Retrieves a single test case. Administrators can retrieve any test case. Normal users can only retrieve if the test case is active and visibility is PUBLIC (otherwise returns 404).",
        "operationId": "getTestCaseById",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "testCaseId",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the test case",
            "schema": {
              "type": "string",
              "example": "65f0123456789abcdef01236"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test case details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "testCase": {
                          "$ref": "#/components/schemas/TestCase"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "patch": {
        "tags": ["Test Cases"],
        "summary": "Update an existing test case",
        "description": "Partially updates an existing test case. Restricted strictly to administrators (ADMIN role). Records administrative audit log.",
        "operationId": "updateTestCase",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "testCaseId",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the test case to update",
            "schema": {
              "type": "string",
              "example": "65f0123456789abcdef01236"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTestCaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Test case updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "testCase": {
                          "$ref": "#/components/schemas/TestCase"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "delete": {
        "tags": ["Test Cases"],
        "summary": "Deactivate a test case",
        "description": "Soft-deactivates an existing test case (`isActive: false`). Restricted strictly to administrators (ADMIN role). Records administrative audit log.",
        "operationId": "deactivateTestCase",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "testCaseId",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the test case to deactivate",
            "schema": {
              "type": "string",
              "example": "65f0123456789abcdef01236"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test case deactivated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Test case deactivated successfully" }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/api/v1/submissions": {
      "post": {
        "tags": ["Submissions"],
        "summary": "Submit code solution (Asynchronous)",
        "description": "Validates submission, persists initial record, and enqueues execution job into Redis/BullMQ for asynchronous processing by dedicated workers. Returns immediately with QUEUED status and PENDING verdict. Fails fast with 422 if problem has zero active test cases. Enforces submission rate limits.",
        "operationId": "createSubmission",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubmissionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Submission created and successfully queued for asynchronous processing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "submission": {
                          "$ref": "#/components/schemas/Submission"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "422": {
            "$ref": "#/components/responses/ProblemNotReadyError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitError"
          }
        }
      }
    },
    "/api/v1/submissions/me": {
      "get": {
        "tags": ["Submissions"],
        "summary": "Get authenticated user submission history",
        "description": "Retrieves submissions belonging strictly to the authenticated user. Omits heavy sourceCode by default for performance unless `includeCode=true` is requested. Supports filtering by problemId, language, and verdict.",
        "operationId": "getMySubmissions",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (minimum 1)",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of submissions per page (1 to 100)",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "problemId",
            "in": "query",
            "description": "Filter submissions by problem ID",
            "schema": {
              "type": "string",
              "example": "65f0123456789abcdef01235"
            }
          },
          {
            "name": "language",
            "in": "query",
            "description": "Filter submissions by programming language",
            "schema": {
              "type": "string",
              "enum": ["CPP", "PYTHON", "JAVASCRIPT"]
            }
          },
          {
            "name": "verdict",
            "in": "query",
            "description": "Filter submissions by execution verdict",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "ACCEPTED",
                "WRONG_ANSWER",
                "COMPILATION_ERROR",
                "RUNTIME_ERROR",
                "TIME_LIMIT_EXCEEDED",
                "MEMORY_LIMIT_EXCEEDED"
              ]
            }
          },
          {
            "name": "includeCode",
            "in": "query",
            "description": "Include sourceCode field in submission objects (default false)",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User submission history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "submissions": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Submission"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/v1/submissions/{submissionId}": {
      "get": {
        "tags": ["Submissions"],
        "summary": "Get submission details by ID",
        "description": "Retrieves current or final submission details including execution status (PENDING, QUEUED, RUNNING, COMPLETED, FAILED), evaluation metrics, and full submitted source code. Used by clients to poll for final execution results after creating an asynchronous submission. Enforces object-level authorization: accessible only to the submission owner or administrators (ADMIN role).",
        "operationId": "getSubmissionById",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "submissionId",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the submission",
            "schema": {
              "type": "string",
              "example": "65f0123456789abcdef01237"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Submission details with source code and evaluation metrics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "submission": {
                          "$ref": "#/components/schemas/Submission"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Provide a valid JWT token issued by POST /api/v1/auth/login"
      }
    },
    "responses": {
      "BadRequestError": {
        "description": "Bad Request — Invalid parameter, malformed payload, or validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Title is required and must not be empty",
              "errorCode": "VALIDATION_ERROR",
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "Title is required and must not be empty"
              },
              "requestId": "req_mu576kgd_5826a9e8"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Unauthorized — Missing, invalid, or expired authentication token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Invalid authorization header format. Expected Bearer <token>",
              "errorCode": "AUTHENTICATION_REQUIRED",
              "error": {
                "code": "AUTHENTICATION_REQUIRED",
                "message": "Invalid authorization header format. Expected Bearer <token>"
              },
              "requestId": "req_mu576kgd_5826a9e8"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Forbidden — Insufficient permissions or object-level authorization denial",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Access denied: insufficient permissions",
              "errorCode": "FORBIDDEN",
              "error": {
                "code": "FORBIDDEN",
                "message": "Access denied: insufficient permissions"
              },
              "requestId": "req_mu576kgd_5826a9e8"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Not Found — Requested resource does not exist or is inactive",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Problem not found",
              "errorCode": "RESOURCE_NOT_FOUND",
              "error": {
                "code": "RESOURCE_NOT_FOUND",
                "message": "Problem not found"
              },
              "requestId": "req_mu576kgd_5826a9e8"
            }
          }
        }
      },
...
      "ProblemNotReadyError": {
        "description": "Unprocessable Entity — Problem has zero active test cases and cannot evaluate submissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "This problem has no active test cases and is not ready for submissions.",
              "errorCode": "PROBLEM_NOT_READY",
              "error": {
                "code": "PROBLEM_NOT_READY",
                "message": "This problem has no active test cases and is not ready for submissions."
              },
              "requestId": "req_mu576kgd_5826a9e8"
            }
          }
        }
      },
      "RateLimitError": {
        "description": "Too Many Requests — Rate limit threshold exceeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": {
                "code": "RATE_LIMIT_EXCEEDED",
                "message": "Too many registration attempts. Please try again in 15 minutes."
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "TestCase": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "65f0123456789abcdef01236"
          },
          "problemId": {
            "type": "string",
            "example": "65f0123456789abcdef01235"
          },
          "input": {
            "type": "string",
            "example": "2\n1 2\n"
          },
          "expectedOutput": {
            "type": "string",
            "example": "3\n"
          },
          "visibility": {
            "type": "string",
            "enum": ["PUBLIC", "HIDDEN"],
            "example": "PUBLIC"
          },
          "order": {
            "type": "integer",
            "minimum": 0,
            "example": 1
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateTestCaseRequest": {
        "type": "object",
        "required": ["input", "expectedOutput", "visibility", "order"],
        "properties": {
          "input": {
            "type": "string",
            "example": "2\n1 2\n"
          },
          "expectedOutput": {
            "type": "string",
            "example": "3\n"
          },
          "visibility": {
            "type": "string",
            "enum": ["PUBLIC", "HIDDEN"],
            "example": "PUBLIC"
          },
          "order": {
            "type": "integer",
            "minimum": 0,
            "example": 1
          }
        }
      },
      "UpdateTestCaseRequest": {
        "type": "object",
        "description": "At least one field must be provided for update.",
        "properties": {
          "input": {
            "type": "string",
            "example": "5\n2 3\n"
          },
          "expectedOutput": {
            "type": "string",
            "example": "5\n"
          },
          "visibility": {
            "type": "string",
            "enum": ["PUBLIC", "HIDDEN"],
            "example": "HIDDEN"
          },
          "order": {
            "type": "integer",
            "minimum": 0,
            "example": 2
          }
        }
      },
      "CreateSubmissionRequest": {
        "type": "object",
        "required": ["problemId", "language", "sourceCode"],
        "properties": {
          "problemId": {
            "type": "string",
            "description": "Valid MongoDB ObjectId of the target problem",
            "example": "65f0123456789abcdef01235"
          },
          "language": {
            "type": "string",
            "enum": ["CPP", "PYTHON", "JAVASCRIPT"],
            "example": "CPP"
          },
          "sourceCode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 65536,
            "example": "#include <iostream>\nusing namespace std;\nint main() {\n    int a, b;\n    if (cin >> a >> b) cout << (a + b) << endl;\n    return 0;\n}"
          }
        }
      },
      "Submission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "65f0123456789abcdef01237"
          },
          "userId": {
            "type": "string",
            "example": "65f0123456789abcdef01234"
          },
          "problemId": {
            "type": "string",
            "example": "65f0123456789abcdef01235"
          },
          "language": {
            "type": "string",
            "enum": ["CPP", "PYTHON", "JAVASCRIPT"],
            "example": "CPP"
          },
          "sourceCode": {
            "type": "string",
            "example": "#include <iostream>..."
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "QUEUED", "RUNNING", "COMPLETED", "FAILED"],
            "example": "QUEUED"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "WRONG_ANSWER",
              "COMPILATION_ERROR",
              "RUNTIME_ERROR",
              "TIME_LIMIT_EXCEEDED",
              "MEMORY_LIMIT_EXCEEDED"
            ],
            "example": "PENDING"
          },
          "runtimeMs": {
            "type": "integer",
            "nullable": true,
            "example": 45
          },
          "memoryKb": {
            "type": "integer",
            "nullable": true,
            "example": null
          },
          "testsPassed": {
            "type": "integer",
            "nullable": true,
            "example": 5
          },
          "totalTests": {
            "type": "integer",
            "nullable": true,
            "example": 5
          },
          "queuedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "failedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "total": {
            "type": "integer",
            "example": 42
          },
          "totalPages": {
            "type": "integer",
            "example": 3
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Resource not found"
          },
          "errorCode": {
            "type": "string",
            "example": "RESOURCE_NOT_FOUND"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "RESOURCE_NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Resource not found"
              }
            }
          },
          "requestId": {
            "type": "string",
            "example": "req_mu576kgd_5826a9e8"
          }
        }
      }
    }
  }
}